Rotor rotation now properly calculated from energy
This commit is from 4 hours+ work, because the rotor
would rotate according to the north/south energy
generation even if it was placed on the east/west side.
The turbine showed the east/west generation of course.
Turns out the energy values needed to be synced between
client and server because minecraft only syncs tile
entities when they're created, and never again until you
tell it to do so. Yay.